type change on set_raw_rc_frame

+++ Sabin-Code/src/msp_frames.h
@@ -22,9 +22,9 @@ struct altitude_frame {
 
 struct set_raw_rc_frame {
 	uint16_t roll;
-	unsigned int pitch;
-	unsigned int yaw;
-	unsigned int throttle;
+	uint16_t pitch;
+	uint16_t yaw;
+	uint16_t throttle;
 };
 
 #endif
